pyproject.toml
[project]
name = "hp-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adrf>=0.1.12",
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-social-oauth2>=3.2.0",
    "drf-spectacular[sidecar]>=0.29.0",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mapradar>=0.1.2",
    "redis>=7.1.0",
    "uuid6>=2025.0.1",
]
